% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifIsoResult.R
\name{print.summary.unifIsoResult}
\alias{print.summary.unifIsoResult}
\title{Print method for summary.unifIsoResult objects}
\usage{
\method{print}{summary.unifIsoResult}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class \code{summary.unifIsoResult}.}

\item{digits}{Integer. Number of significant digits to print.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
Invisibly returns the original object \code{x}.
}
\description{
Print method for summary.unifIsoResult objects
}
\examples{
\dontrun{
# TODO: Add basic print.summary example here
}
}
