% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_isotropy.R
\name{uni_iso_grid}
\alias{uni_iso_grid}
\alias{uni_iso_grid.matrix}
\alias{uni_iso_grid.array}
\alias{uni_iso_grid.geodata}
\alias{uni_iso_grid.SpatialGridDataFrame}
\alias{uni_iso_grid.default}
\title{Isotropy test for random fields on a d-dimensional grid}
\usage{
uni_iso_grid(
  spdata,
  objfunc = "normdiff",
  minimize = FALSE,
  lags_list = NULL,
  window_dims = NULL,
  overlap_dims = NULL
)

\method{uni_iso_grid}{matrix}(
  spdata,
  objfunc = "normdiff",
  minimize = FALSE,
  lags_list = NULL,
  window_dims = NULL,
  overlap_dims = NULL
)

\method{uni_iso_grid}{array}(
  spdata,
  objfunc = "normdiff",
  minimize = FALSE,
  lags_list = NULL,
  window_dims = NULL,
  overlap_dims = NULL
)

\method{uni_iso_grid}{geodata}(
  spdata,
  objfunc = "normdiff",
  minimize = FALSE,
  lags_list = NULL,
  window_dims = NULL,
  overlap_dims = NULL
)

\method{uni_iso_grid}{SpatialGridDataFrame}(
  spdata,
  objfunc = "normdiff",
  minimize = FALSE,
  lags_list = NULL,
  window_dims = NULL,
  overlap_dims = NULL
)

\method{uni_iso_grid}{default}(
  spdata,
  objfunc = "normdiff",
  minimize = FALSE,
  lags_list = NULL,
  window_dims = NULL,
  overlap_dims = NULL
)
}
\arguments{
\item{spdata}{A spatial data object containing the random field observed on a rectangular d-dimensional grid.
Can be a `matrix`, `array`, a `SpatialGridDataFrame`, or a `geodata` object.
In the case of a matrix or array with scalar entries, where the last dimension
represents the coordinate index of a multivariate field. The array may also contain
general objects (lists).}

\item{objfunc}{Character string or function. The objective function to be evaluated
over directions to obtain the argmax. Must be vectorized. Default is `"normdiff"`, which
computes the squared Euclidean norm of the difference and is implemented in C++ for efficiency.
This works only for scalar or multivariate numeric data.}

\item{minimize}{Logical. Should the objective function be minimized instead of maximized?
Default is `FALSE` (maximize).}

\item{lags_list}{A list of matrices, where each matrix represents a set of lag vectors
(each row is a lag vector in the scale of the grid step size). If not specified,
defaults to a single lag set of unit vectors in each spatial dimension and their negatives.
Lag vectors in the same matrix need to be of the same norm and fulfill the condition
outlined in Baumhakel et al. (2026+).}

\item{window_dims}{Integer vector. Dimensions of the moving window used to estimate
the asymptotic long-run covariance matrix. If not specified, defaults to a vector of 2's.}

\item{overlap_dims}{Integer vector. Dimensions of overlap for the moving window. If not specified,
defaults to a vector of zeros.}
}
\value{
An object of class `unifIsoResult` containing:
  \item{statistic}{The overall test statistic value.}
  \item{stat_vec}{Vector of test statistics for each lag vector.}
  \item{p_value}{The p-value of the test.}
  \item{moments}{List containing the estimated proportions of maximizers and long-run covariance matrix (`sig`).}
  \item{eigenvalues}{Eigenvalues of the estimated covariance matrix.}
  \item{n_obs}{Vector of the number of observations used for each lag set (edge of the domain cannot be used).}
  \item{lags_list}{The used list of lags.}
  \item{window_dims}{The used window dimensions.}
  \item{overlap_dims}{The used overlap dimensions.}
}
\description{
Performs a uniformity-based isotropy test for spatial data on a regular grid based on
Baumhakel et al. (2026+).
The function accepts matrices, multidimensional arrays, `SpatialGridDataFrame`, and `geodata` objects.
Depending on the data type, it evaluates the given objective function over specified spatial lags
at every point in the field to find maximizing directions. The proportion of maximizers
are compared to the uniform distribution. to determine whether the field exhibits directional dependence (anisotropy).
}
\examples{
set.seed(1)

# Only run this example if geoR is installed
if (requireNamespace("geoR", quietly = TRUE)) {

  # simple example ----------------------------------------------------------
  # 2D domain, univariate, anisotropic
  sim1 <- geoR::grf(grid = expand.grid(1:30, 1:30),
                    cov.pars = c(1, 1), aniso.pars = c(0,2))
  res1 <- uni_iso_grid(sim1, window_dims = c(8, 8))
  # all proportions lie outside joint confidence intervals
  summary(res1)
  plot(res1)


  # custom lag set, multivariate --------------------------------------------
  sim2 <- geoR::grf(grid = expand.grid(1:30, 1:30), nsim = 2,
                    cov.pars = c(1, 2), aniso.pars = c(0,2))

  # test 2 sets of lags simultaneously
  lags_list = list(rbind(diag(2, 2), diag(-2, 2)),
                   matrix(c(1,1,-1,-1,1,-1,1,-1), nrow = 4, ncol = 2))

  res2 <- uni_iso_grid(sim2, lags_list = lags_list,
                       window_dims = c(8, 8))
  # anisotropy is not detected in all directions
  summary(res2)
  plot(res2)

  # + custom objective function (sum of absolute differences, vectorized)
  f2 <- function(z0, zlagged) {
    diffs <- zlagged - matrix(rep(z0, nrow(zlagged)), nrow = nrow(zlagged), byrow = TRUE)
    return(rowSums(abs(diffs)))
  }
  res2a <- uni_iso_grid(sim2, objfunc = f2, lags_list = lags_list,
                        window_dims = c(8, 8))


  # 3D domain, list-valued --------------------------------------------------

  # generate exemplary list-valued field
  sim3 <- geoR::grf(grid = expand.grid(1:10, 1:10, 1:10), nsim = 2,
                    cov.pars = c(1, 0.25))
  dat3 <- sim3$dat
  dat3 <- ifelse(dat3 > 0, "+", "-")
  sim3.l <- lapply(1:nrow(dat3), function(i) dat3[i,])
  dim(sim3.l) <- c(10, 10, 10)

  # custom objective function (x is single observation, y is list of observations)
  f3 <- function(x, yl) {
    return(sapply(yl, function(y) {sum(unlist(x) != unlist(y))}))
  }

  res3 <- uni_iso_grid(sim3.l, objfunc = f3)
  res3
  plot(res3)

}
}
\references{
Baumhakel, J., Hörmann, S., & Neumann M. (2026+). \emph{Testing isotropy of random fields with values in abstract spaces}
}
