% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unisono-package.R
\docType{package}
\name{unisono-package}
\alias{unisono}
\alias{unisono-package}
\title{unisono: UNiformity-based ISOtropy testing for Nonscalar Objects}
\description{
Performs isotropy test for random fields observed on regular grids. Supports multivariate and abstract data. Provides efficient C++ implementations to evaluate directional dependencies across spatial lags using standard or user-defined objective functions. Includes diagnostic visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baumhakel/unisono}
  \item Report bugs at \url{https://github.com/baumhakel/unisono/issues}
}

}
\author{
\strong{Maintainer}: Julius Baumhakel \email{baumhakel@tugraz.at}

Authors:
\itemize{
  \item Julius Baumhakel \email{baumhakel@tugraz.at}
  \item Siegfried Hörmann
  \item Matthias Neumann
}

}
\keyword{internal}
